modernize FE, remove sockets, fix BE for add somethins               concept: Dashboard -> for Graphs using chart.js               concept: User Orders -> Need to write API

+++ backend/src/main/java/cz/syntaxbro/erpsystem/models/InventoryItem.java
@@ -35,7 +35,7 @@ public class InventoryItem {
 
     @PositiveOrZero(message = "Item quantity must be zero or positive")
     @Column(name = "stocked_amount")
-    private int stockedAmount;
+    private double stockedAmount;
 
     @CreationTimestamp
     private LocalDateTime createdAt;
